rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "camelot-py[cv]>=1.0.9",
    "dotenv>=0.9.9",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-teddynote>=0.5.1",
    "langgraph>=0.6.7",
    "numpy>=2.2.6",
    "openai>=1.107.1",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "pi-heif>=1.1.0",
    "psycopg[binary]>=3.2.10",
    "pymupdf4llm>=0.0.27",
    "streamlit>=1.49.1",
    "tabula-py>=2.10.0",
    "tqdm>=4.67.1",
    "unstructured>=0.18.14",
    "unstructured-inference>=1.0.5",
]
